Fix/throttles on register

+++ backend/src/user/user.service.ts
@@ -438,6 +438,14 @@ export class UserService {
     }
 
     async deleteUser(id: string) {
+		const user = await this.prisma.user.findUnique({
+            where: { id },
+        });
+
+        if (!user) {
+            throw new NotFoundException('User not found');
+        }
+		
         return await this.prisma.user.delete({
             where: { id },
         });
